🧹 [code health improvement] Extract setup steps to a separate file
